Add files via upload
This is the program that i have created. It is like a game whice we play now or in our childhood with our friends "Stone-Paper-Scissor". The name of the game is "Snake-Water-Gun". This is a game between a player and a computer that take the input from the user and computer itself generates its own input.